backend/michicondrias_core/requirements.txt
fastapi[all]>=0.111.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9

uvicorn[standard]>=0.30.0
mangum>=0.17.0
boto3>=1.34.0
